feat: move some of the gateway types and functional trait to primitives
