+++ gateway/pallets/runtime-gateway/src/lib.rs
@@ -46,6 +46,8 @@ mod tests;
 
 pub use pallet::*;
 
+// TODO: probably move this stuff to primitives
+
 // Bits of transfer flags associated with types on-gateway of transfer.
 bitflags! {
 
@@ -75,7 +77,6 @@ bitflags! {
 
 // Bits of storage flags associated with types on-gateway of transfer.
 bitflags! {
-
     /// Flags used by a contract to customize calls.
     #[derive(Encode, Decode)]
     pub struct CallFlags: u8 {
